+++ tests/test_dataframe/test_get.py
@@ -302,6 +302,14 @@ def test_get_slice():
     with pytest.raises(RuntimeError):
         df.get_slice(2, 4)
 
+    # empty DataFrame
+    df = rc.DataFrame(sort=True)
+    assert_frame_equal(df.get_slice(3, 3), rc.DataFrame(sort=True))
+
+    df = rc.DataFrame(sort=True, columns=['a', 'b'])
+    assert_frame_equal(df.get_slice(3, 3), rc.DataFrame(sort=True, columns=['a', 'b']))
+
+    # full DataFrame
     df = rc.DataFrame({'a': [1, 2, 3, 4], 'b': [5, 6, 7, 8]}, columns=['a', 'b'], index=[2, 4, 6, 8], sort=True)
 
     assert_frame_equal(df.get_slice(2, 8), df)